Move HelpFormatter to FastlaneCore
