+++ fastlane_core/lib/fastlane_core/ui/help.erb
