+++ scan/lib/scan/commands_generator.rb
@@ -1,7 +1,7 @@
 require 'commander'
 
 require 'fastlane_core/configuration/configuration'
-require 'fastlane/commands_generator/help_formatter'
+require 'fastlane_core/ui/help_formatter'
 require_relative 'module'
 require_relative 'manager'
 require_relative 'options'
@@ -29,7 +29,7 @@ module Scan
       program :help, "Author", "Felix Krause <scan@krausefx.com>"
       program :help, "Website", "https://fastlane.tools"
       program :help, "Documentation", "https://docs.fastlane.tools/actions/scan/"
-      program :help_formatter, Fastlane::CommandsGenerator::HelpFormatter
+      program :help_formatter, FastlaneCore::HelpFormatter
 
       global_option("--verbose") { FastlaneCore::Globals.verbose = true }
